[demo] update zoom-i demo tile
